+++ src/compiler/bytecodebuilder.c
@@ -7,7 +7,7 @@
 #include "compiler/compile.h"
 #include "compiler/bytecodebuilder.h"
 
-void emit_ins_abc(lfCompilerCtx *ctx, lfOpCode op, uint8_t a, uint8_t b, uint8_t c, size_t lineno) {
+void emit_ins_abc(lfCompilerCtx *ctx, lfOpCode op, uint8_t a, uint8_t b, uint8_t c, int lineno) {
     array_push(&ctx->current, op);
     array_push(&ctx->current, a);
     array_push(&ctx->current, b);
@@ -15,7 +15,7 @@ void emit_ins_abc(lfCompilerCtx *ctx, lfOpCode op, uint8_t a, uint8_t b, uint8_t
     array_push(&ctx->linenumbers, lineno);
 }
 
-void emit_ins_ad(lfCompilerCtx *ctx, lfOpCode op, uint8_t a, uint16_t d, size_t lineno) {
+void emit_ins_ad(lfCompilerCtx *ctx, lfOpCode op, uint8_t a, uint16_t d, int lineno) {
     array_push(&ctx->current, op);
     array_push(&ctx->current, a);
     array_push(&ctx->current, (d >> 0) & 0xFF);
@@ -23,20 +23,20 @@ void emit_ins_ad(lfCompilerCtx *ctx, lfOpCode op, uint8_t a, uint16_t d, size_t 
     array_push(&ctx->linenumbers, lineno);
 }
 
-void emit_ins_e(lfCompilerCtx *ctx, lfOpCode op, uint32_t e, size_t lineno) {
+void emit_ins_e(lfCompilerCtx *ctx, lfOpCode op, uint32_t e, int lineno) {
     array_push(&ctx->current, op);
     emit_u24(ctx, e);
     array_push(&ctx->linenumbers, lineno);
 }
 
-void emit_ins_e_at(lfCompilerCtx *ctx, lfOpCode op, uint32_t e, size_t idx, size_t lineno) {
-    size_t curr = length(&ctx->current);
+void emit_ins_e_at(lfCompilerCtx *ctx, lfOpCode op, uint32_t e, int idx, int lineno) {
+    int curr = length(&ctx->current);
     length(&ctx->current) = idx;
     emit_ins_e(ctx, op, e, lineno);
     length(&ctx->current) = curr;
 }
 
-void emit_op(lfCompilerCtx *ctx, lfOpCode op, size_t lineno) {
+void emit_op(lfCompilerCtx *ctx, lfOpCode op, int lineno) {
     emit_ins_e(ctx, op, 0, lineno);
 }
 
@@ -51,8 +51,8 @@ uint32_t new_u64(lfCompilerCtx *ctx, uint64_t value) {
     return length(&ctx->ints) - 1;
 }
 
-uint32_t new_string(lfCompilerCtx *ctx, char *value, size_t length) {
-    for (size_t i = 0; i < length(&ctx->strings); i++) { /* TODO: Replace with a map for O(1) insertion */
+uint32_t new_string(lfCompilerCtx *ctx, char *value, int length) {
+    for (int i = 0; i < length(&ctx->strings); i++) { /* TODO: Replace with a map for O(1) insertion */
         if (!strncmp(ctx->strings[i], value, length))
             return i;
     }
