+++ src/compiler/variablemap.c
@@ -20,13 +20,13 @@ void lf_variablebucket_deleter(lfVariableBucket **bucket) {
     }
 }
 
-size_t lf_variablemap_compute_hash(const char *s) {
+int lf_variablemap_compute_hash(const char *s) {
     int p = 59;
     int m = 1e9 + 9;
-    size_t hash_value = 0;
-    size_t p_pow = 1;
-    size_t len = strlen(s);
-    for (size_t i = 0; i < len; i++) {
+    int hash_value = 0;
+    int p_pow = 1;
+    int len = strlen(s);
+    for (int i = 0; i < len; i++) {
         char c = s[i];
         if (c >= 'a' && c <= 'z') hash_value = (hash_value + (s[i] - 'a' + 1) * p_pow) % m;
         else if (c >= 'A' && c <= 'Z') hash_value = (hash_value + (s[i] - 'A' + 27) * p_pow) % m;
@@ -37,19 +37,19 @@ size_t lf_variablemap_compute_hash(const char *s) {
     return hash_value;
 }
 
-lfVariableMap lf_variablemap_create(size_t size) {
+lfVariableMap lf_variablemap_create(int size) {
     lfVariableMap map = array_new(lfVariableBucket *, lf_variablebucket_deleter);
     array_reserve(&map, size);
-    for (size_t i = 0; i < size; i++)
+    for (int i = 0; i < size; i++)
         array_push(&map, NULL);
     return map;
 }
 
 lfVariableMap lf_variablemap_clone(const lfVariableMap *map) {
-    size_t map_size = length(map);
+    int map_size = length(map);
     lfVariableMap new_map = lf_variablemap_create(map_size);
 
-    for (size_t i = 0; i < map_size; ++i) {
+    for (int i = 0; i < map_size; ++i) {
         lfVariableBucket *current = (*map)[i];
         lfVariableBucket *prev_new_bucket = NULL;
 
@@ -75,7 +75,7 @@ lfVariableMap lf_variablemap_clone(const lfVariableMap *map) {
 }
 
 bool lf_variablemap_lookup(const lfVariableMap *map, const char *key, lfVariable *out) {
-    size_t hash = lf_variablemap_compute_hash(key) % length(map);
+    int hash = lf_variablemap_compute_hash(key) % length(map);
     lfVariableBucket *b = (*map)[hash];
     if (b == NULL) return false;
     if (b->next == NULL) {
@@ -93,7 +93,7 @@ bool lf_variablemap_lookup(const lfVariableMap *map, const char *key, lfVariable
 }
 
 void lf_variablemap_insert(lfVariableMap *map, const char *key, lfVariable value) {
-    size_t hash = lf_variablemap_compute_hash(key) % length(map);
+    int hash = lf_variablemap_compute_hash(key) % length(map);
     lfVariableBucket *next = malloc(sizeof(lfVariableBucket));
     next->key = key;
     next->value = value;
